+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to the Reactodia will be documented in this document.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+#### 🐛 Fixed
+- Embed the web fonts used by the diagram into an exported image, as it is not able to load an external font and displayed the text with a fallback font instead. Can be tuned or disabled with the `embedFonts` export option. (by [@aindlq](https://github.com/aindlq))
+
 
 ## [0.35.2] - 2026-08-08
 #### 🐛 Fixed